+++ src/components/icons/X.tsx
@@ -1,0 +1,33 @@
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+import React from 'react'
+
+import { Icon, IconProps } from '@chakra-ui/react'
+
+export const XIcon = (props: IconProps) => (
+  <Icon {...props}>
+    <svg
+      width='17'
+      height='16'
+      viewBox='0 0 17 16'
+      fill='none'
+      xmlns='http://www.w3.org/2000/svg'
+    >
+      <g clipPath='url(#clip0_703_2142)'>
+        <path
+          d='M10.3552 6.77143L16.3115 0H14.9001L9.72817 5.87954L5.59738 0H0.833008L7.07957 8.8909L0.833008 15.9918H2.24455L7.70622 9.78279L12.0686 15.9918H16.833L10.3548 6.77143H10.3552ZM8.42187 8.96923L7.78897 8.0839L2.75315 1.03921H4.92121L8.98517 6.7245L9.61807 7.60983L14.9007 14.9998H12.7327L8.42187 8.96957V8.96923Z'
+          fill='#5C5959'
+        />
+      </g>
+      <defs>
+        <clipPath id='clip0_703_2142'>
+          <rect
+            width='16'
+            height='16'
+            fill='white'
+            transform='translate(0.833008)'
+          />
+        </clipPath>
+      </defs>
+    </svg>
+  </Icon>
+)